fix(host): close review gaps
